Cloud Functions/Extracting Stops/requirements.txt
functions-framework==3.*
requests
google-cloud-bigquery
google-cloud-secret-manager